% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameutils.R
\name{bind_pp_df}
\alias{bind_pp_df}
\title{Bind PP dfs together}
\usage{
bind_pp_df(this_team_short, pp_5v4, pp_5v3, pp_4v3)
}
\arguments{
\item{this_team_short}{String of 3-letter short team name.}

\item{pp_5v4}{Vector scraped from GS}

\item{pp_5v3}{Vector scraped from GS}

\item{pp_4v3}{Vector scraped from GS}
}
\value{
Data frame
}
\description{
Bind PP dfs together
}

