% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baseutils.R
\name{clean_team_short}
\alias{clean_team_short}
\title{Convert 3-ltter team name to our standard 3-letter team_short}
\usage{
clean_team_short(team_short)
}
\arguments{
\item{team_short}{3-letter team name}
}
\value{
corrected (if necessary) team_short.
   Unchanged, unless team is T.B, N.J, S.J, or L.A.
}
\description{
Convert 3-ltter team name to our standard 3-letter team_short
}
\examples{
clean_team_short( "T.B" )
}

