% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseutils.R
\name{clock_time_to_start_cum}
\alias{clock_time_to_start_cum}
\title{Convert scoreboard (period, clock time) to cumulative time}
\usage{
clock_time_to_start_cum(period, clock_time, session_id = "2", round = 3)
}
\arguments{
\item{period}{Numeral representing period.}

\item{session_id}{A string, "2" or "3" to indicate regular season or playoffs.
Defaults to "2".}

\item{clocktime}{String with MM:SS format}
}
\value{
A numeric value denoting cumulative time in min
from beginning of game.
}
\description{
Convert scoreboard (period, clock time) to cumulative time
}

