% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gameutils.R
\name{get_game_id10}
\alias{get_game_id10}
\title{Convert (season, session_id, game_id4) triplet into game_id10}
\usage{
get_game_id10(season, session_id, game_id4)
}
\arguments{
\item{season}{String of form "20152016"}

\item{session_id}{String of "2" or "3"}

\item{game_id4}{String like "0097"}
}
\description{
Convert (season, session_id, game_id4) triplet into game_id10
}

