% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gameutils.R
\name{get_gc_name_id}
\alias{get_gc_name_id}
\title{internal function. Extract NHL ID and name from GC Roster row.}
\usage{
get_gc_name_id(gc_row)
}
\arguments{
\item{gc_row}{A row from GC roster.}
}
\value{
(NHL ID, last_name, first_name, position)
Don't export.
}
\description{
internal function. Extract NHL ID and name from GC Roster row.
}

