% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameutils.R
\name{get_gc_rosters}
\alias{get_gc_rosters}
\title{Scrape roster from GameCenter Live site for a specific game}
\usage{
get_gc_rosters(season = "20152016", session_id = "2", game_id4 = "0001")
}
\arguments{
\item{season}{String of form "20152016"}

\item{session_id}{String of session ID, "2" or "3"}

\item{game_id4}{String of 4-digit game ID, e.g., "0037"}
}
\value{
Data frame of rosters for both teams
}
\description{
Scrape roster from GameCenter Live site for a specific game
}

