% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameutils.R
\name{get_gs_url}
\alias{get_gs_url}
\title{Get Game Summary url}
\usage{
get_gs_url(season, session_id, game_id4)
}
\arguments{
\item{season}{String, "20152016"}

\item{session_id}{String "2" or "3"}

\item{game_id4}{String "0001"}
}
\value{
string
}
\description{
Get Game Summary url
}
\examples{
get_gs_url( "20152016", "2", "0001" )
}

