% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gameutils.R
\name{get_ha_number}
\alias{get_ha_number}
\title{Convert (ha, number) pair into ha_number}
\usage{
get_ha_number(ha, number)
}
\arguments{
\item{ha}{String "H" or "A"}

\item{number}{Jersey number}
}
\value{
String of form ha_number, unique to a player in a game.
}
\description{
Convert (ha, number) pair into ha_number
}

