% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameutils.R
\name{get_pp_results}
\alias{get_pp_results}
\title{Get PP results data frame from a single game}
\usage{
get_pp_results(season, session_id, game_id4, team_tbl)
}
\arguments{
\item{season}{String for season, e.g., "20152016"}

\item{session_id}{String for session id, "2" or "3"}

\item{game_id4}{String of 4-digit game ID, e.g.,  "0003"}

\item{team_tbl}{Collected team_tbl}
}
\value{
Data frame with PP results for this game
}
\description{
Get PP results data frame from a single game
}

