% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameutils.R
\name{get_pp_results_df}
\alias{get_pp_results_df}
\title{Get df of PP results from GS scraped PP vector.}
\usage{
get_pp_results_df(pp_results_vector, strength = "5v4")
}
\arguments{
\item{pp_results_vector}{Vector scraped from GS of form ( "1-3", "02:12" )}

\item{strength}{string}
}
\value{
Data frame of PP results from one strength scenario
}
\description{
Get df of PP results from GS scraped PP vector.
}

