% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameutils.R
\name{get_roster_from_list}
\alias{get_roster_from_list}
\title{internal function. Get roster from GameCenter Live site}
\usage{
get_roster_from_list(homeaway, raw_list)
}
\arguments{
\item{back}{Days back from today (Sys.Date()).  0 is today, -1 is yestday.}

\item{game_date}{String in YYYMMDD form.
use game_date if both back and game_date are specified.}
}
\value{
Data frame of season, session_id, game_id4, for game_date games.
Don't export.
}
\description{
internal function. Get roster from GameCenter Live site
}

