% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameutils.R
\name{get_season_end}
\alias{get_season_end}
\title{Get Season end YYYY from season string}
\usage{
get_season_end(season_str)
}
\arguments{
\item{season_str}{String in form YYYYYYYY, e.g., "20152016"}
}
\value{
Numeric YYYY of season end, 2016 in above example
}
\description{
Get Season end YYYY from season string
}

