% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseutils.R
\name{height_to_inches}
\alias{height_to_inches}
\title{Height to inches.}
\usage{
height_to_inches(ht)
}
\arguments{
\item{ht}{String of height in form 5'11" or 5'11.}
}
\value{
Integer of height in inches
}
\description{
Height to inches.
}
\examples{
height_to_inches( "5'11" )
}

