% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlutils.R
\docType{package}
\name{nhlutils-package}
\alias{nhlutils}
\alias{nhlutils-package}
\title{nhlutils: A package for working with data scraped from NHL.com}
\description{
The nhlutils package provides functions to handle the
problems that frequently emerge when working with NHL data.
}
\details{
baseutils:
 These are simple functions to deal with conversions of time
 and name formats.

gameutils:
 working with NHL-specific game data, involving their standard formats
 for season, session, game IDs.

Details
}

