% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nhlutils.R
\docType{package}
\name{nhlutils}
\alias{nhlutils}
\alias{nhlutils-package}
\title{nhlutils: A package for working with data scraped from NHL.com}
\description{
The nhlutils package provides functions to handle the
problems that commonly emerge when working with NHL data.
}
\section{baseutils}{

 These are simple functions to deal with frequently occuring
 conversions of time and name formats.
}

