% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseutils.R
\name{simple_cap}
\alias{simple_cap}
\title{Capitalize the first letter of each word (usually names)}
\usage{
simple_cap(x)
}
\arguments{
\item{x}{A name, first and last separated by space}
}
\value{
String with first letter capitalized
}
\description{
Capitalize the first letter of each word (usually names)
}
\examples{
simple_cap( "ben smith" )
}

