% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gameutils.R
\name{split_game_id10}
\alias{split_game_id10}
\title{Convert game_id10 into (season, session_id, game_id4)}
\usage{
split_game_id10(game_id10)
}
\arguments{
\item{game_id10}{10-character game id, e.g., "2015020037".
Note that the embedded year is the season *start* year.  This
might be unexpected behavior but this is what GC Live does.}
}
\value{
Data frame with season (8-character), session_id, game_id4
}
\description{
Convert game_id10 into (season, session_id, game_id4)
}

