% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baseutils.R
\name{time_decimal_to_mmss}
\alias{time_decimal_to_mmss}
\title{Convert x.y decimal time to MM:SS.}
\usage{
time_decimal_to_mmss(time_decimal)
}
\arguments{
\item{time_decimal}{A numerical value of time.}
}
\value{
a string of time
}
\description{
Convert x.y decimal time to MM:SS.
}

