% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseutils.R
\name{time_mmss_to_decimal}
\alias{time_mmss_to_decimal}
\title{Convert MM:SS time to MM + SS/60.}
\usage{
time_mmss_to_decimal(mmss, digits = 3)
}
\arguments{
\item{mmss}{A time in MM:SS format}
}
\value{
Rounded numeric value of decimal time.
}
\description{
Convert MM:SS time to MM + SS/60.
}
\examples{
time_mmss_to_decimal( "5:16" )
}

