% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameutils.R
\name{get_game_id10}
\alias{get_game_id10}
\title{Convert (season, session_id, game_id4) triplet into game_id10}
\usage{
get_game_id10(season, session_id, game_id4)
}
\arguments{
\item{season}{String of form "20152016"}

\item{session_id}{String of "2" or "3"}

\item{game_id4}{String like "0097"}
}
\value{
10-char game ID.  NOTE: uses season *start* ("2015" of "20152016").
This is GC convention, not mine.
}
\description{
Convert (season, session_id, game_id4) triplet into game_id10
}

