% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrutils.R
\name{get_season_standings}
\alias{get_season_standings}
\title{Get Season standings from hockey-reference.com}
\usage{
get_season_standings(season, team_tbl = NULL)
}
\arguments{
\item{season}{String for season, e.g., "20152016"}

\item{team_tbl}{Team tbl from nhl_db}
}
\value{
data frame of standings
}
\description{
Get Season standings from hockey-reference.com
}

